requirements.txt
RPi.GPIO
python-mpd2
